Add new test case bug1
Because of a bug in the compiler, this test will fail while outputting
the cfg since a deleted instruction is used as an argument. This is
because the value that is assigned to the variable is a common
subexpression, so it is deleted. Since it is an assignment in a loop,
it is used in a phi operand. Right now CSE won't update the arguments of
prior instructions, as in the case of a backward branch in a while loop.