Fix subtle bug in live range calculation
This fix will ensure that if a phi in the same block is used as an
argument for another phi, that phi (the previous value) will be live
when the block is entered. This is somewhat confusing, but it is
necessary since the semantics of phi instructions is that they all occur
in parallel at the same point in time. With this fix, the test case for
phi move order dependency (phi2) passes.